==> code/markdown/01c_impute_autosomal.Rmd <==
---
title: "Impute probes with detectionP > 0.01"
subtitle: "Look at autosomal probes"
author: 
  - David lukacsovich^[University of Miami]
  - Lily Wang^[University of Miami]
date: "`r Sys.Date()`"
output:
  rmarkdown::html_document:
    highlight: pygments
    theme: lumen
    toc: true
    number_sections: true
    df_print: paged
    code_download: false
    toc_float:
      collapsed: yes
    toc_depth: 3
editor_options:
  chunk_output_type: inline    
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Set Up

## Load Libraries

```{r library}
suppressMessages({
  library(dplyr)
  library(methyLImp2)
})

base_dir <- "../../.."
analysis_dir <- file.path(base_dir, "AD_CR")
data_dir <- file.path(analysis_dir, "data/DNAm_Aplus/processed")
probe_dir <- file.path(analysis_dir, "data/DNAm_Aplus/probes")
```

# Load Data

## Load Data

```{r load_probes}
dnam_data <- readRDS(file.path(data_dir, "Preprocessed_Beta.RDS"))
pheno_df <- dnam_data %>%
  SummarizedExperiment::colData() %>%
  as.data.frame()

probe_df <- dnam_data %>%
  SummarizedExperiment::rowData() %>%
  as.data.frame() %>%
  dplyr::rename(cpg = "Name") %>%
  dplyr::filter(!(.data$seqnames %in% c("chrM", "chrX", "chrY")))
dnam_data <- dnam_data[probe_df$cpg, ]
```

## Load Target Probes

```{r load_detp}
detp_df <- read.csv(file.path(probe_dir, "detectionP_Failures.csv")) %>%
  dplyr::filter(.data$cpg %in% probe_df$cpg) %>%
  dplyr::filter(.data$sample %in% pheno_df$sample) %>%
  dplyr::filter(.data$detp >= 0.01)
```

# Impute Together

## Define Functions

```{r func_detp}
get_beta_data <- function(dnam_data, detp_df) {
  beta_data <- SummarizedExperiment::assays(dnam_data)$dnam %>%
    t()
  
  samples <- row.names(beta_data)
  probes <- colnames(beta_data)
  
  detp_df <- detp_df %>%
    dplyr::filter(.data$sample %in% samples) %>%
    dplyr::filter(.data$cpg %in% probes)
  
  for (Sample in unique(detp_df$sample)) {
    cpgs <- detp_df %>%
      dplyr::filter(.data$sample == Sample) %>%
      dplyr::pull("cpg")
    
    beta_data[Sample, cpgs] <- NA
  }
  
  beta_data
}

get_probe_df <- function(dnam_data) {
  probe_df <- dnam_data %>%
    SummarizedExperiment::rowData() %>%
    as.data.frame() %>%
    dplyr::select(cpg = "Name", chr = "seqnames")
  
  probe_df
}

get_batch_summaries <- function(detp_df, pheno_df) {
  missing_df <- detp_df %>%
    dplyr::select("cpg", "batch") %>%
    dplyr::group_by(.data$batch, .data$cpg) %>%
    dplyr::summarise(missing_count = n()) %>%
    dplyr::ungroup()
  
  count_df <- pheno_df %>%
    dplyr::group_by(.data$batch) %>%
    dplyr::summarise(total_count = n()) %>%
    dplyr::ungroup()
  
  batch_df <- missing_df %>%
    dplyr::left_join(count_df, by = "batch") %>%
    dplyr::mutate(
      present_count = .data$total_count - .data$missing_count,
      label = paste0(.data$batch, "_", .data$cpg)
    ) %>%
    dplyr::select(
      "cpg", "batch", "label", "total_count", "present_count", "missing_count"
    )
  
  batch_df
}

get_na_targets <- function(pheno_df, detp_df) {
  detp_df <- detp_df %>%
    dplyr::left_join(
      pheno_df %>% dplyr::select("sample", "batch"),
      by = "sample"
    ) %>%
    dplyr::select("cpg", "sample", "batch") %>%
    dplyr::mutate(label = paste0(.data$batch, "_", .data$cpg))
  
  print(dim(detp_df))
  
  batch_df <- get_batch_summaries(detp_df, pheno_df) %>%
    dplyr::filter(.data$present_count < 10)
  
  detp_df <- detp_df %>%
    dplyr::filter(.data$label %in% batch_df$label) %>%
    dplyr::select("cpg", "sample")
  
  print(dim(detp_df))
  
  detp_df
}

impute_beta_nobatch <- function(beta_data, probe_df) {
  samples <- row.names(beta_data)
  probes <- colnames(beta_data)
  
  beta_data <- methyLImp2::methyLImp2(
    beta_data,
    type = "user",
    annotation = probe_df,
    BPPARAM = BiocParallel::SnowParam(workers = 22, exportglobals = FALSE),
    groups = NULL,
    overwrite_res = FALSE,
    minibatch_frac = 0.2,
    minibatch_reps = 3
  )
  
  colnames(beta_data) <- probes
  row.names(beta_data) <- samples
  
  beta_data
}
```

## Filter Detection P

```{r filter_detp}
pheno_df <- dnam_data %>%
  SummarizedExperiment::colData() %>%
  as.data.frame()

detp_df <- get_na_targets(pheno_df, detp_df)
```

## Run Impute

```{r impute_run_full}
beta_data <- get_beta_data(dnam_data, detp_df)
probe_df <- get_probe_df(dnam_data)

beta_data <- impute_beta_nobatch(beta_data, probe_df)
dnam_data <- SummarizedExperiment::SummarizedExperiment(
  assays = list(dnam = t(beta_data)),
  rowData = dnam_data %>%
    SummarizedExperiment::rowData() %>%
    as.data.frame(),
  colData = dnam_data %>%
    SummarizedExperiment::colData() %>%
    as.data.frame()
)

saveRDS(dnam_data, file.path(data_dir, "Imputed_Autosomal.RDS"))
```

# Session Information

<details>
  <summary>**Session Info**</summary>
```{r session}
sessionInfo()
```
</details>
