==> code/markdown/02g_get_miami_ad_statistics.Rmd <==
---
title: "Compare significant probes to Miami-AD results"
author: 
  - David lukacsovich^[University of Miami]
  - Lily Wang^[University of Miami]
date: "`r Sys.Date()`"
output:
  rmarkdown::html_document:
    highlight: pygments
    theme: lumen
    toc: true
    number_sections: true
    df_print: paged
    code_download: false
    toc_float:
      collapsed: yes
    toc_depth: 3
editor_options:
  chunk_output_type: inline    
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Set Up

## Load Libraries

```{r library}
suppressMessages({
  library(dplyr)
  library(readxl)
})

base_dir <- "../../.."
analysis_dir <- file.path(base_dir, "AD_CR")
probe_dir <- file.path(analysis_dir, "data/DNAm_Aplus/probes")
data_dir <- file.path(analysis_dir, "data/combp")
reference_dir <- file.path(analysis_dir, "data/ref_files")
save_dir <- file.path(
  analysis_dir, "analysis_results/only_amyloid_positive_2_stage_approach"
)
```

# Define Functions

## Get Miami-AD Data

```{r func_miami}
load_miami_ad <- function(label) {
  fname <- file.path(
    save_dir, "comb-p-DMRs", paste0(label, "_MIAMI_AD.xlsx")
  )
  
  dataset_df <- readxl::read_excel(fname, sheet = "Dataset Abbreviations") %>%
    dplyr::mutate(
      PMID = unlist(lapply(stringr::str_split(.data$PMID, "/"), `[`, 4))
    )
  cpg_df <- readxl::read_excel(fname, sheet = "Individual Datasets") %>%
    dplyr::rename(cohort = "sample_group")
  
  miami_df <- cpg_df %>%
    dplyr::left_join(dataset_df, by = c("dataset" = "Dataset")) %>%
    dplyr::filter(.data$pValue < 0.05) %>%
    dplyr::rename(pValue_miami = "pValue")
  
  miami_df
}
```

## Get Annotations

```{r func_annot}
load_probe_annotations <- function(label) {
  save_name <- file.path(
    save_dir,
    paste0("linear_model_", label),
    "PHC_stats_bacon_interact.csv"
  )
  
  res_df <- read.csv(save_name)
  
  res_df <- res_df %>%
    dplyr::mutate(
      direction = ifelse(.data$estimate_bacon > 0, "+", "-")
    ) %>%
    dplyr::select(
      "CpG",
      "Chr",
      "Position",
      "GREAT_annotation",
      estimate = "estimate_bacon",
      stdErr = "std_error_bacon",
      pValue_study = "pvalue_bacon",
      fdr = "fdr_bacon",
      Direction = "direction"
    )
  
  res_df
}
```

## Get Target Functions

```{r func_probes}
get_key_probes <- function(label) {
  fname <- file.path(
    save_dir, "comb-p-DMRs", paste0(label, "_mQTLs_inputs.csv")
  )
  
  cpg_df <- read.csv(fname) %>%
    dplyr::rename(CpG = "cpgs", Source = "Sources")
  
  cpg_df
}
```

## Mange Compiled Data

```{r func_compile}
load_full_data <- function(label) {
  miami_df <- load_miami_ad(label)
  annot_df <- load_probe_annotations(label)
  probe_df <- get_key_probes(label)
  
  cpg_df <- annot_df %>%
    dplyr::inner_join(probe_df, by = "CpG") %>%
    dplyr::inner_join(miami_df, by = "CpG")
}

save_full_data <- function(label, cpg_df) {
  fname <- file.path(
    save_dir, "comb-p-DMRs", paste0(label, "_MIAMI_AD_Results.csv")
  )
  
  write.csv(cpg_df, fname, row.names = FALSE)
}

analysis_wrapper <- function(label) {
  cpg_df <- load_full_data(label)
  save_full_data(label, cpg_df)
}
```

# Compile Results

```{r run}
analysis_wrapper("CN")
analysis_wrapper("MCI")
```

# Session Information

<details>
  <summary>**Session Info**</summary>
```{r session}
sessionInfo()
```
</details>
