==> code/markdown/06a_mem_correlation.Rmd <==
---
title: "Get correlation between PHC_MEM and DNAm_residual or PHC_pTau"
author: 
  - David lukacsovich^[University of Miami]
  - Lily Wang^[University of Miami]
date: "`r Sys.Date()`"
output:
  rmarkdown::html_document:
    highlight: pygments
    theme: lumen
    toc: true
    number_sections: true
    df_print: paged
    code_download: false
    toc_float:
      collapsed: yes
    toc_depth: 3
editor_options:
  chunk_output_type: inline    
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Set Up

## Load Libraries

```{r library}
suppressMessages({
  library(dplyr)
  library(SummarizedExperiment)
  library(parallel)
  library(doParallel)
})

base_dir <- "../../.."
analysis_dir <- file.path(base_dir, "AD_CR")
data_dir <- file.path(analysis_dir, "data/DNAm_Aplus/processed")
reference_dir <- file.path(analysis_dir, "data/ref_files")
save_dir <- file.path(
  analysis_dir, "analysis_results/only_amyloid_positive_2_stage_approach"
)
```

# Define Functions

## Get Residuals

```{r func_resid}
run_resid_test <- function(pheno_df, fo) {
  formula <- as.formula(fo)
  lm_mod <- stats::lm(formula, data = pheno_df)
  resids <- stats::residuals(lm_mod)
  
  resids
}

calculate_residuals <- function(pheno_df) {
  fo <- paste(
    c(
      "PHC_MEM ~ ",
      "age_at_visit",
      "Gender",
      "APOE4",
      "B",
      "NK",
      "CD4T",
      "Mono",
      "Gran"
    ),
  collapse = " + "
  )
  
  resid_vals <- run_resid_test(pheno_df, fo)
  
  pheno_df$resid <- resid_vals
  
  pheno_df
}

add_residuals <- function(dnam_data) {
  pheno_df <- as.data.frame(SummarizedExperiment::colData(dnam_data)) %>%
    dplyr::mutate(
      PHC_MEM = as.numeric(.data$PHC_MEM),
      age_at_visit = as.numeric(.data$age_at_visit),
      Gender = as.character(.data$Gender),
      APOE4 = as.numeric(.data$APOE4),
      DX = as.character(.data$DX),
      B = as.numeric(.data$B),
      NK = as.numeric(.data$NK),
      CD4T = as.numeric(.data$CD4T),
      CD8T = as.numeric(.data$CD8T),
      Mono = as.numeric(.data$Mono),
      Neutro = as.numeric(.data$Neutro),
      Eosino = as.numeric(.data$Eosino)
    ) %>%
    dplyr::mutate(
      Gran = .data$Neutro + .data$Eosino
    )
  
  pheno_df <- calculate_residuals(pheno_df)
  
  dnam_data$resid <- pheno_df$resid
  
  dnam_data
}
```

## Get Load

```{r func_load}
filter_probes <- function(dnam_data) {
  row_df <- dnam_data %>%
    SummarizedExperiment::rowData() %>%
    as.data.frame() %>%
    dplyr::filter(
      !.data$filter_crosshyb,
      !.data$filter_snp
    )
  
  dnam_data <- dnam_data[row_df %>% dplyr::pull("Name"),]
}

get_dnam_data <- function() {
  dnam_data <- readRDS(file.path(data_dir, "Batched_PCA.RDS"))
  dnam_data <- filter_probes(dnam_data)
  dnam_data <- dnam_data[,!is.na(dnam_data$PHC_Diagnosis)]
  dnam_data <- dnam_data[,!is.na(dnam_data$APOE4)]
  dnam_data <- add_residuals(dnam_data)
  
  dnam_data
}
```


## Parallelize

```{r func_parallel}
start_parallel <- function(parallel, cores) {
  if (parallel &&
      requireNamespace("doParallel", quietly = TRUE) &&
      requireNamespace("parallel", quietly = TRUE)) {
    if (Sys.info()["sysname"] == "Windows"){
      cluster <- parallel::makeCluster(cores)
      doParallel::registerDoParallel(cluster)
    } else {
      doParallel::registerDoParallel(cores)
      cluster <- NULL
    }
  } else {
    parallel = FALSE
    cluster <- NULL
  }

  list(parallel = parallel, cluster = cluster)
}

stop_parallel <- function(parallel, cluster) {
  if (parallel &&
      requireNamespace("doParallel", quietly = TRUE) &&
      requireNamespace("parallel", quietly = TRUE)) {
    doParallel::stopImplicitCluster()
    if (is.null(cluster) && Sys.info()["sysname"] == "Windows") {
      parallel::stopCluster(cluster)
    }
  }

  TRUE
}
```

## Annotations

```{r func_annot}
get_nasser_data <- function() {
  celltypes <- readxl::read_xlsx(
    file.path(reference_dir, "Nassser study selected biosamples.xlsx"),
    col_names = FALSE
  ) %>%
    dplyr::pull(1)
  df <- readr::read_tsv(
    file.path(
      reference_dir, 
      "AllPredictions.AvgHiC.ABC0.015.minus150.ForABCPaperV3.txt.gz"
    )
  ) %>%
    dplyr::filter(.data$CellType %in% celltypes) %>%
    dplyr::filter(!.data$isSelfPromoter)  %>% 
    dplyr::filter(.data$class != "promoter") %>%
    dplyr::select(
      seqnames = "chr",
      "start",
      "end",
      celltype = "CellType"
    ) %>%
    dplyr::distinct()
  
  df
}

get_hg19_positions <- function() {
  minfi_object <- utils::data(
    "IlluminaHumanMethylationEPICanno.ilm10b4.hg19",
    package = "IlluminaHumanMethylationEPICanno.ilm10b4.hg19"
  )
  anno_df <- minfi::getAnnotation(minfi_object) %>%
    as.data.frame() %>%
    dplyr::select(
      CpG = "Name",
      seqnames = "chr",
      start = "pos",
      end = "pos"
    )
  
  anno_df
}

get_nasser_annotation <- function() {
  nasser_df <- get_nasser_data() %>%
    dplyr::mutate(
      start = .data$start - 250,
      end = .data$end + 250
    )
  anno_df <- get_hg19_positions()
  
  nasser_gr <- nasser_df %>%
    GenomicRanges::makeGRangesFromDataFrame()
  anno_gr <- anno_df %>%
    GenomicRanges::makeGRangesFromDataFrame()
  
  hits_df <- findOverlaps(anno_gr, nasser_gr) %>%
    as.data.frame()
  
  anno_df <- data.frame(
    CpG = anno_df$CpG[hits_df$queryHits],
    CellType = nasser_df$celltype[hits_df$subjectHits]
  ) %>%
    dplyr::distinct() %>%
    dplyr::group_by(.data$CpG) %>%
    dplyr::summarise(CellType = paste0(.data$CellType, collapse = ";")) %>%
    dplyr::ungroup()
  
  anno_df
}

get_stored_annotations <- function(dnam_data) {
  row_df <- dnam_data %>%
    SummarizedExperiment::rowData() %>%
    as.data.frame() %>%
    dplyr::mutate(
      CrossHybridize = as.integer(.data$filter_crosshyb),
      nearSNP = as.integer(.data$filter_snp)
    ) %>%
    dplyr::select(
      CpG = "Name",
      Chr = "seqnames",
      Position = "pos",
      "UCSC_RefGene_Name",
      "UCSC_RefGene_Group",
      "Relation_to_Island"
    )
  
  row_df
}

get_great_annotations <- function() {
  great_df <- readRDS(file.path(reference_dir, "GREAT_EPIC.RDS")) %>%
    dplyr::select(
      CpG = "cpg", GREAT_annotation
    )
  
  great_df
}

get_probe_annotations <- function(dnam_data) {
  nasser_df <- get_nasser_annotation()
  row_df <- get_stored_annotations(dnam_data)
  great_df <- get_great_annotations()
  
  probe_df <- row_df %>%
    dplyr::left_join(great_df, by = "CpG") %>%
    dplyr::left_join(nasser_df, by = "CpG") %>%
    dplyr::mutate(
      nasser_is_enhancer = ifelse(is.na(.data$CellType), 0, 1)
    ) %>%
    tidyr::replace_na(list(CellType = "")) %>%
    dplyr::select(
      "CpG",
      "Chr",
      "Position",
      "GREAT_annotation",
      "UCSC_RefGene_Name",
      "UCSC_RefGene_Group",
      "Relation_to_Island",
      nasser_enhancer_cell_type = "CellType",
      "nasser_is_enhancer"
    )
  
  probe_df
}
```

## Association Test

```{r func_assoc}
run_lm_step <- function(beta_value, pheno_df, formula) {
  data <- data.frame(dnam = beta_value, pheno_df)
  lm_mod <- stats::lm(formula, data = data)
  
  coef_tbl <- data.frame(summary(lm_mod)$coefficients)
  df_resid <- nrow(data) - length(stats::coef(lm_mod))

  t_vals <- coef_tbl[, "t.value"]
  p_vals <- coef_tbl[, "Pr...t.."]
  z_vals = stats::qnorm(stats::pt(t_vals, df_resid))
  
  coef_df <- janitor::clean_names(coef_tbl) %>%
    dplyr::mutate(
      pvalue = p_vals,
      zvalue = z_vals,
      df = df_resid
    )
  
  dnam_df <- coef_df["dnam",] %>%
    dplyr::select(
      "df",
      estimate_dnam = "estimate",
      std_error_dnam = "std_error",
      tvalue_dnam = "t_value",
      pvalue_dnam = "pvalue",
      zvalue_dnam = "zvalue"
    )
  tau_df <- coef_df["PHC_pTau",] %>%
    dplyr::select(
      estimate_ptau = "estimate",
      std_error_ptau = "std_error",
      tvalue_ptau = "t_value",
      pvalue_ptau = "pvalue",
      zvalue_ptau = "zvalue"
    )
  inter_df <- coef_df["PHC_pTau:dnam",] %>%
    dplyr::select(
      estimate_interact = "estimate",
      std_error_interact = "std_error",
      tvalue_interact = "t_value",
      pvalue_interact = "pvalue",
      zvalue_interact = "zvalue"
    )
  
  coef_df <- cbind(dnam_df, tau_df, inter_df)

  coef_df
}

run_lm_test <- function(beta_vals, pheno_df, fo) {
  parallel_res <- start_parallel(TRUE, 8)
  do_parallel <- parallel_res$parallel
  cluster <- parallel_res$cluster
  
  formula <- stats::as.formula(fo)
  
  results <- plyr::adply(
    beta_vals,
    .margins = 1,
    .fun = function(beta_value) {
      run_lm_step(beta_value, pheno_df, formula)
    }, .parallel = do_parallel
  )
  
  stop_parallel(do_parallel, cluster)
  
  results <- results %>%
    dplyr::rename(probe = "X1") %>%
    dplyr::mutate(
      fdr_dnam = stats::p.adjust(.data$pvalue_dnam, method = "fdr"),
      fdr_ptau = stats::p.adjust(.data$pvalue_ptau, method = "fdr"),
      fdr_interact = stats::p.adjust(.data$pvalue_interact, method = "fdr")
    )
  
  results
}
```

## Wrapper Function

```{r func_wrapper}
run_association <- function(dnam_data) {
  beta_vals <- SummarizedExperiment::assays(dnam_data)$dnam
  beta_vals <- Harman::shiftBetas(beta_vals, shiftBy = 1e-4)
  pheno_df <- dnam_data %>%
    SummarizedExperiment::colData() %>%
    as.data.frame()
  
  fo <- "resid ~ PHC_pTau + dnam + dnam:PHC_pTau + PHC_Education"

  run_lm_test(beta_vals, pheno_df, fo)
}

get_association_statistics <- function(dnam_data) {
  stats_df <- run_association(dnam_data)
  annot_df <- get_probe_annotations(dnam_data)
  stats_df <- annot_df %>%
    dplyr::right_join(stats_df, by = c("CpG" = "probe"))
  
  stats_df
}

run_model_analysis <- function(
    dnam_data,
    keep = c(1, 2, 3),
    label = ""
) {
  dnam_data <- dnam_data[,dnam_data$PHC_Diagnosis %in% keep]
  pheno_df <- dnam_data %>%
    SummarizedExperiment::colData() %>%
    as.data.frame()
  stats_df <- get_association_statistics(dnam_data)
  
  save_name <- "PHC_stats.csv"
  
  save_dir <- file.path(
    save_dir, 
    paste0("linear_model_", label, "_education")
  )
  
  if (!dir.exists(save_dir)) {
    dir.create(save_dir)
  }
  
  write.csv(
    stats_df,
    file = file.path(save_dir, save_name),
    row.names = FALSE
  )
  
  write.csv(
    pheno_df,
    file = file.path(save_dir, "Phenotypes.csv"),
    row.names = FALSE
  )

}
```

# Run Analysis

## Get Data

```{r load}
dnam_data <- get_dnam_data()
```

## Cognitive Normal

```{r run_cn}
run_model_analysis(
  dnam_data, keep = c(1), label = "CN"
)
```

## MCI

```{r run_mci}
run_model_analysis(
  dnam_data, keep = c(2), label = "MCI"
)
```

## AD

```{r run_ad}
run_model_analysis(
  dnam_data, keep = c(3), label = "AD"
)
```

# Session Information

<details>
  <summary>**Session Info**</summary>
```{r session}
sessionInfo()
```
</details>
