==> code/markdown/06b_bacon_correct.Rmd <==
---
title: "Run Bacon Correction"
author: 
  - David lukacsovich^[University of Miami]
  - Lily Wang^[University of Miami]
date: "`r Sys.Date()`"
output:
  rmarkdown::html_document:
    highlight: pygments
    theme: lumen
    toc: true
    number_sections: true
    df_print: paged
    code_download: false
    toc_float:
      collapsed: yes
    toc_depth: 3
editor_options:
  chunk_output_type: inline    
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Set Up

## Load Libraries

```{r library}
suppressMessages({
  library(dplyr)
  library(SummarizedExperiment)
  library(parallel)
  library(doParallel)
})

base_dir <- "../../.."
analysis_dir <- file.path(base_dir, "AD_CR")
data_dir <- file.path(analysis_dir, "data/DNAm_Aplus/processed")
reference_dir <- file.path(analysis_dir, "data/ref_files")
save_dir <- file.path(
  analysis_dir, "analysis_results/only_amyloid_positive_2_stage_approach"
)
```

# Define Functions

## Load Data

```{r func_load}
get_stat_data <- function(label, save_dir) {
  save_name <- "PHC_stats.csv"
  
  stats_df <- read.csv(file.path(save_dir, save_name))
  
  columns <- colnames(stats_df)
  meta_columns <- columns[
    (!stringr::str_starts(columns, "estimate")) &
      (!stringr::str_starts(columns, "std_error")) &
      (!stringr::str_starts(columns, "tvalue")) &
      (!stringr::str_starts(columns, "pvalue")) &
      (!stringr::str_starts(columns, "zvalue")) &
      (!stringr::str_starts(columns, "fdr"))
  ]
  
  label_columns <- c("estimate", "std_error", "tvalue", "pvalue", "zvalue", "fdr")
  select_columns <- c(meta_columns, paste(label_columns, label, sep = "_"))
  
  stats_df <- stats_df[,select_columns]
  colnames(stats_df) <- c(meta_columns, label_columns)
  
  stats_df
}

get_dnam_data <- function() {
  dnam_data <- readRDS(file.path(data_dir, "Batched_PCA.RDS"))
  dnam_data <- dnam_data[,!is.na(dnam_data$PHC_Diagnosis)]
  dnam_data <- dnam_data[,!is.na(dnam_data$APOE4)]
  
  dnam_data
}
```

## Get IQR

```{r func_iqr}
get_iqr <- function(dnam_data, keep = c(1, 2, 3)) {
  dnam_data <- dnam_data[,dnam_data$PHC_Diagnosis %in% keep]
  beta_vals <- SummarizedExperiment::assays(dnam_data)$dnam
  
  iqr_df <- data.frame(
    CpG = row.names(beta_vals),
    IQR = unname(apply(beta_vals * 100, MARGIN = 1, stats::IQR))
  )
  
  iqr_df
}

add_iqr_fdr <- function(stats_df, min_iqr = 2) {
  stats_df$fdr_exclude_iqr_lt2 <- NA
  stats_df$fdr_exclude_iqr_lt2_bacon <- NA
  stats_df$fdr_alt_bacon <- NA
  
  targets <- stats_df$IQR >= min_iqr
  
  stats_df$fdr_exclude_iqr_lt2[targets] <- stats::p.adjust(
    stats_df$pvalue[targets], method = "fdr"
  )
  stats_df$fdr_exclude_iqr_lt2_bacon[targets] <- stats::p.adjust(
    stats_df$pvalue_bacon[targets], method = "fdr"
  )
  
  stats_df$fdr_alt_bacon[!targets] <- stats::p.adjust(
    stats_df$pvalue_bacon[!targets], method = "fdr"
  )
  
  stats_df
}
```

## Parallelize

```{r func_parallel}
start_parallel <- function(parallel, cores) {
  if (parallel &&
      requireNamespace("doParallel", quietly = TRUE) &&
      requireNamespace("parallel", quietly = TRUE)) {
    if (Sys.info()["sysname"] == "Windows"){
      cluster <- parallel::makeCluster(cores)
      doParallel::registerDoParallel(cluster)
    } else {
      doParallel::registerDoParallel(cores)
      cluster <- NULL
    }
  } else {
    parallel = FALSE
    cluster <- NULL
  }

  list(parallel = parallel, cluster = cluster)
}

stop_parallel <- function(parallel, cluster) {
  if (parallel &&
      requireNamespace("doParallel", quietly = TRUE) &&
      requireNamespace("parallel", quietly = TRUE)) {
    doParallel::stopImplicitCluster()
    if (is.null(cluster) && Sys.info()["sysname"] == "Windows") {
      parallel::stopCluster(cluster)
    }
  }

  TRUE
}
```

## Bacon Correction

```{r func_bacon}
run_bacon_correction <- function(
    data,
    alpha = 1.28,
    beta = 0.36,
    lambda1 = 0,
    lambda2 = 3,
    tau1 = 1000,
    tau2 = 100,
    gamma1 = 99,
    gamma2 = 0.5
) {
  set.seed(23)
  bc <- bacon::bacon(
    teststatistics = NULL,
    effectsizes =  data$estimate,
    standarderrors = data$std_error,
    na.exclude = TRUE,
    priors = list(
      sigma = list(alpha = alpha,  beta = beta), 
      mu = list(lambda = c(lambda1, lambda2, -lambda2), tau = c(tau1, tau2, tau2)), 
      epsilon = list(gamma = c(gamma1, gamma2, gamma2))
    )
  )
  
  tvalue <- data$estimate / data$std_error
  zvalue <- stats::qnorm(stats::pt(tvalue, data$df))
  chisq <- zvalue ^ 2
  inflation_base <- median(chisq,na.rm = TRUE) / qchisq(0.5, 1)
  inflation_bacon <- bacon::inflation(bc)
  bias_bacon <- bacon::bias(bc)
  
  inflation_effect <- c(
    inflation_base = as.numeric(inflation_base),
    inflation_bacon = as.numeric(inflation_bacon),
    bias_bacon = as.numeric(bias_bacon)
  )
  
  data_bacon <- data %>%
    dplyr::mutate(
      estimate_bacon = bacon::es(bc),
      std_error_bacon = bacon::se(bc),
      pvalue_bacon = bacon::pval(bc)
    ) %>%
    dplyr::mutate(
      fdr_bacon = stats::p.adjust(.data$pvalue_bacon, method = "fdr")
    )

  set.seed(42)
  bc2 <- bacon::bacon(
    teststatistics = NULL,
    effectsizes = data_bacon$estimate_bacon,
    standarderrors = data_bacon$std_error_bacon,
    na.exclude = TRUE,
    priors = list(
      sigma = list(alpha = alpha,  beta = beta), 
      mu = list(lambda = c(lambda1, lambda2, -lambda2), tau = c(tau1, tau2, tau2)), 
      epsilon = list(gamma = c(gamma1, gamma2, gamma2))
    )
  )
  
  tvalue <- data_bacon$estimate_bacon / data_bacon$std_error_bacon
  zvalue <- stats::qnorm(stats::pt(tvalue, data$df))
  chisq <- zvalue ^ 2
  inflation_correction <- median(chisq,na.rm = TRUE)/qchisq(0.5, 1)
  inflation_bacon_correction <- bacon::inflation(bc2)
  bias_bacon_correction <- bacon::bias(bc2)
  
  inflation_effect <- c(
    inflation_original = as.numeric(inflation_base),
    inflation_bacon = as.numeric(inflation_bacon),
    bias_bacon = as.numeric(bias_bacon),
    inflation_correction = as.numeric(inflation_correction),
    inflation_bacon_correction = as.numeric(inflation_bacon_correction),
    bias_bacon_correction = as.numeric(bias_bacon_correction)
  )
  
  list(data = data_bacon, inflation_effect = inflation_effect)
}

evaluate_bacon_correction <- function(
    data,
    alpha = 1.28,
    beta = 0.36,
    lambda1 = 0,
    lambda2 = 3,
    tau1 = 1000,
    tau2 = 100,
    gamma1 = 99,
    gamma2 = 0.5
) {
  res_ls <- run_bacon_correction(
    data,
    alpha = alpha,
    beta = beta,
    lambda1 = lambda1,
    lambda2 = lambda2,
    tau1 = tau1,
    tau2 = tau2,
    gamma1 = gamma1,
    gamma2 = gamma2
  )
  
  inflation_effect <- res_ls$inflation_effect
  
  data.frame(
    alpha = alpha,
    beta = beta,
    lambda1 = lambda1,
    lambda2 = lambda2,
    tau1 = tau1,
    tau2 = tau2,
    gamma1 = gamma1,
    gamma2 = gamma2,
    inflation = inflation_effect[["inflation_correction"]],
    inflation_bacon = inflation_effect[["inflation_bacon_correction"]],
    bias_bacon = inflation_effect[["bias_bacon_correction"]]
  )
}
```

## Iterarive Bacon

```{r func_iter}
get_random_parameters <- function(count = 100, seed = 12) {
  set.seed(seed)
  data.frame(
    step = 1:count,
    alpha = round(runif(count, min = 0.5, max = 2), digits = 3),
    beta = round(runif(count, min = 0.1, max = 1), digits = 3),
    lambda1 = round(runif(count, min = -1, max = 1), digits = 3),
    lambda2 = round(runif(count, min = 1, max = 4), digits = 3),
    tau1 = round(10**runif(count, min = 2, max = 4), digits = 1),
    tau2 = runif(count, min = 0.02, max = 0.5),
    gamma1 = round(runif(count, min = 80, max = 99), digits = 1),
    gamma2 = runif(count, min = 0.1, max = 0.9)
  ) %>%
    dplyr::mutate(
      tau2 = round(.data$tau1 * .data$tau2, digits = 1),
      gamma2 = round((100 - .data$gamma1) * .data$gamma2, digits = 3)
    )
}

run_bacon_iter <- function(data, random_df) {
  parallel_res <- start_parallel(TRUE, 16)
  do_parallel <- parallel_res$parallel
  cluster <- parallel_res$cluster
  
  res_df <- plyr::adply(
    .data = random_df,
    .margins = 1,
    .fun = function(row) {
      evaluate_bacon_correction(
        data,
        alpha = row$alpha,
        beta = row$beta,
        lambda1 = row$lambda1,
        lambda2 = row$lambda2,
        tau1 = row$tau1,
        tau2 = row$tau2,
        gamma1 = row$gamma1,
        gamma2 = row$gamma2
      )
    }, .parallel = do_parallel
  )
  
  stop_parallel(do_parallel, cluster)
  
  res_df
}
```

## Bacon Wrapper

```{r func_bacon_wrapper}
bacon_analyze <- function(data) {
  random_df <- get_random_parameters(count = 1024, seed = 24)
  iter_df <- run_bacon_iter(data, random_df)
  iter_df <- iter_df %>%
  dplyr::mutate(
    score1 = abs(1 - .data$inflation_bacon),
    score2 = abs(.data$bias_bacon),
    score = pmax(.data$score1, .data$score2)
  ) %>%
    dplyr::arrange(.data$score)
  
  row <- iter_df[1,]
  
  res <- run_bacon_correction(
    data,
    alpha = row$alpha,
    beta = row$beta,
    lambda1 = row$lambda1,
    lambda2 = row$lambda2,
    tau1 = row$tau1,
    tau2 = row$tau2,
    gamma1 = row$gamma1,
    gamma2 = row$gamma2
    
  )
  inflation_effect <- res$inflation_effect
  data <- res$data
  
  inflation_df <- data.frame(
    labels = names(inflation_effect)[1:3],
    pre_correction = as.numeric(inflation_effect)[1:3],
    post_correction = as.numeric(inflation_effect)[4:6]
  )
  
  list(stats_df = data, inflation_df = inflation_df)
}

bacon_wrapper <- function(label, iqr_df, save_label = "") {
  save_dir <- file.path(
    save_dir,
    paste0("linear_model_", save_label, "_education")
  )
  stats_df <- get_stat_data(label, save_dir)
  
  res_ls <- bacon_analyze(stats_df)
  inflation_df <- res_ls$inflation_df
  stats_df <- res_ls$stats_df
  
  stats_df <- stats_df %>%
    dplyr::left_join(iqr_df, by = "CpG")
  
  stats_df <- add_iqr_fdr(stats_df, min_iqr = 2)
  
  save_name <- paste0("PHC_stats_bacon_", label, ".csv")
  inflation_name <- paste0("PHC_inflation_", label, ".csv")
  
  write.csv(
    inflation_df,
    file = file.path(save_dir, inflation_name),
    row.names = FALSE
  )

  write.csv(
    stats_df,
    file = file.path(save_dir, save_name),
    row.names = FALSE
  )
}
```

# Get IQR Information

```{r iqr}
dnam_data <- get_dnam_data()
cn_df <- get_iqr(dnam_data, keep = 1)
mci_df <- get_iqr(dnam_data, keep = 2)
ad_df <- get_iqr(dnam_data, keep = 3)
```

# Run Bacon Correction

## Cognitive Normal

```{r bacon_cn}
bacon_wrapper("dnam", cn_df, save_label = "CN")
bacon_wrapper("ptau", cn_df, save_label = "CN")
bacon_wrapper("interact", cn_df, save_label = "CN")
```

## MCI

```{r bacon_mci}
bacon_wrapper("dnam", mci_df, save_label = "MCI")
bacon_wrapper("ptau", mci_df, save_label = "MCI")
bacon_wrapper("interact", mci_df, save_label = "MCI")
```

## AD

```{r bacon_ad}
bacon_wrapper("dnam", ad_df, save_label = "AD")
bacon_wrapper("ptau", ad_df, save_label = "AD")
bacon_wrapper("interact", ad_df, save_label = "AD")
```

# Session Information

<details>
  <summary>**Session Info**</summary>
```{r session}
sessionInfo()
```
</details>
